Remove an import cycle
