+++ datafiles/models.py
@@ -1,25 +1,12 @@
 import dataclasses
-from dataclasses import dataclass
-from typing import Dict, Optional
 
 import log
 from classproperties import classproperty
 
 from . import hooks
-from .converters import Converter, map_type
-from .managers import InstanceManager, ModelManager
-
-
-@dataclass
-class ModelMeta:
-    datafile_attrs: Optional[Dict[str, Converter]] = None
-    datafile_pattern: Optional[str] = None
-
-    datafile_manual: bool = False
-    datafile_defaults: bool = False
-    datafile_auto_load: bool = True
-    datafile_auto_save: bool = True
-    datafile_auto_attr: bool = False
+from .builders import build_datafile
+from .managers import ModelManager
+from .meta import ModelMeta
 
 
 class Model:
@@ -44,49 +31,13 @@ class Model:
                 elif path:
                     self.datafile.save()
 
-                hooks.apply(self, self.datafile, build_datafile)
+                hooks.apply(self, self.datafile)
 
         log.debug(f'Initialized {self.__class__} object')
 
     @classproperty
     def datafiles(cls) -> ModelManager:  # pylint: disable=no-self-argument
         return ModelManager(cls)
-
-
-def build_datafile(obj, root=None) -> InstanceManager:
-    try:
-        return object.__getattribute__(obj, 'datafile')
-    except AttributeError:
-        log.debug(f"Building 'datafile' for {obj.__class__} object")
-
-    m = getattr(obj, 'Meta', None)
-    pattern = getattr(m, 'datafile_pattern', None)
-    attrs = getattr(m, 'datafile_attrs', None)
-    manual = getattr(m, 'datafile_manual', ModelMeta.datafile_manual)
-    defaults = getattr(m, 'datafile_defaults', ModelMeta.datafile_defaults)
-    auto_load = getattr(m, 'datafile_auto_load', ModelMeta.datafile_auto_load)
-    auto_save = getattr(m, 'datafile_auto_save', ModelMeta.datafile_auto_save)
-    auto_attr = getattr(m, 'datafile_auto_attr', ModelMeta.datafile_auto_attr)
-
-    if attrs is None and dataclasses.is_dataclass(obj):
-        attrs = {}
-        log.debug(f'Mapping attributes for {obj.__class__} object')
-        for field in dataclasses.fields(obj):
-            self_name = f'self.{field.name}'
-            if pattern is None or self_name not in pattern:
-                attrs[field.name] = map_type(field.type, name=field.name)
-
-    return InstanceManager(
-        obj,
-        attrs=attrs,
-        pattern=pattern,
-        manual=manual,
-        defaults=defaults,
-        auto_load=auto_load,
-        auto_save=auto_save,
-        auto_attr=auto_attr,
-        root=root,
-    )
 
 
 def create_model(
